Move identifier-case report helper into identifier-case module
Relocate src/plugin/src/reporting/identifier-case-report.js to src/plugin/src/identifier-case/identifier-case-report.js because the module only serves identifier-case plan/context helpers. Keeping it with the identifier-case utilities clarifies the domain boundary and removes the orphaned reporting directory.
