+++ src/plugin/tests/identifier-case-locals.integration.test.js
@@ -13,7 +13,7 @@ import {
     clearIdentifierCaseDryRunContexts
 } from "../src/identifier-case/identifier-case-context.js";
 import { prepareIdentifierCasePlan } from "../src/identifier-case/local-plan.js";
-import { maybeReportIdentifierCaseDryRun } from "../src/reporting/identifier-case-report.js";
+import { maybeReportIdentifierCaseDryRun } from "../src/identifier-case/identifier-case-report.js";
 
 const currentDirectory = fileURLToPath(new URL(".", import.meta.url));
 const pluginPath = path.resolve(currentDirectory, "../src/gml.js");
