+++ src/plugin/tests/identifier-case-report.test.js
@@ -11,7 +11,7 @@ import {
     setIdentifierCaseDryRunContext,
     clearIdentifierCaseDryRunContexts
 } from "../src/identifier-case/identifier-case-context.js";
-import { maybeReportIdentifierCaseDryRun } from "../src/reporting/identifier-case-report.js";
+import { maybeReportIdentifierCaseDryRun } from "../src/identifier-case/identifier-case-report.js";
 
 const currentDirectory = fileURLToPath(new URL(".", import.meta.url));
 const pluginPath = path.resolve(currentDirectory, "../src/gml.js");
